+++ src/test/java/de/codescape/jira/plugins/scrumpoker/model/GlobalSettingsTest.java
@@ -14,7 +14,7 @@ public class GlobalSettingsTest {
         GlobalSettings globalSettings = new GlobalSettings();
         assertThat(globalSettings.getSessionTimeout(), is(equalTo(SESSION_TIMEOUT_DEFAULT)));
         assertThat(globalSettings.getAllowRevealDeck(), is(equalTo(ALLOW_REVEAL_DECK_DEFAULT)));
-        assertThat(globalSettings.isDefaultProjectActivation(), is(equalTo(DEFAULT_PROJECT_ACTIVATION_DEFAULT)));
+        assertThat(globalSettings.isActivateScrumPoker(), is(equalTo(ACTIVATE_SCRUM_POKER_DEFAULT)));
         assertThat(globalSettings.isDisplayDropdownOnBoards(), is(equalTo(DISPLAY_DROPDOWN_ON_BOARDS_DEFAULT)));
         assertThat(globalSettings.isCheckPermissionToSaveEstimate(), is(equalTo(CHECK_PERMISSION_TO_SAVE_ESTIMATE_DEFAULT)));
         assertThat(globalSettings.getDisplayCommentsForIssue(), is(equalTo(DISPLAY_COMMENTS_FOR_ISSUE_DEFAULT)));