+++ docs/changelog.md
@@ -14,6 +14,7 @@ title: Changelog
 * refactor: rename variables in health check to better explain checks
 * docs: add documentation to logic for need to talk flag
 * style: clean up formatting of plugin configuration
+* feat: rename global setting for Scrum Poker activation (#105)
 
 ### [20.05.2] - 2020-05-14